Move general test setup code from test_install.py to testing.py.
